refactor: filter abstraction
