+++ frontend/src/component/feature/FeatureToggleList/FeatureToggleFilters/FeatureToggleFilters.tsx
@@ -1,55 +1,14 @@
 import { useEffect, useState, VFC } from 'react';
-import { Box, styled } from '@mui/material';
 import useProjects from 'hooks/api/getters/useProjects/useProjects';
-import { ConditionallyRender } from 'component/common/ConditionallyRender/ConditionallyRender';
-import AddFilterButton from './AddFilterButton/AddFilterButton';
 import { useSegments } from 'hooks/api/getters/useSegments/useSegments';
-import { FilterDateItem } from 'component/common/FilterDateItem/FilterDateItem';
-import {
-    FilterItem,
-    FilterItemParams,
-} from 'component/common/FilterItem/FilterItem';
 import useAllTags from 'hooks/api/getters/useAllTags/useAllTags';
+import { FilterItemParamHolder, Filters, IFilterItem } from '../../../filter/Filters';
 
-const StyledBox = styled(Box)(({ theme }) => ({
-    display: 'flex',
-    padding: theme.spacing(2, 3),
-    gap: theme.spacing(1),
-    flexWrap: 'wrap',
-}));
-
-type FeatureTogglesListFilters = {
-    project?: FilterItemParams | null | undefined;
-    tag?: FilterItemParams | null | undefined;
-    state?: FilterItemParams | null | undefined;
-    segment?: FilterItemParams | null | undefined;
-    createdAt?: FilterItemParams | null | undefined;
-};
 
 interface IFeatureToggleFiltersProps {
-    state: FeatureTogglesListFilters;
-    onChange: (value: FeatureTogglesListFilters) => void;
+    state: FilterItemParamHolder;
+    onChange: (value: FilterItemParamHolder) => void;
 }
-
-type IBaseFilterItem = {
-    label: string;
-    options: {
-        label: string;
-        value: string;
-    }[];
-    filterKey: keyof FeatureTogglesListFilters;
-};
-
-type ITextFilterItem = IBaseFilterItem & {
-    singularOperators: [string, ...string[]];
-    pluralOperators: [string, ...string[]];
-};
-
-type IDateFilterItem = IBaseFilterItem & {
-    dateOperators: [string, ...string[]];
-};
-
-type IFilterItem = ITextFilterItem | IDateFilterItem;
 
 export const FeatureToggleFilters: VFC<IFeatureToggleFiltersProps> = ({
     state,
@@ -71,31 +30,6 @@ export const FeatureToggleFilters: VFC<IFeatureToggleFiltersProps> = ({
     ];
 
     const [availableFilters, setAvailableFilters] = useState<IFilterItem[]>([]);
-    const [unselectedFilters, setUnselectedFilters] = useState<string[]>([]);
-    const [selectedFilters, setSelectedFilters] = useState<string[]>([]);
-
-    const deselectFilter = (label: string) => {
-        const newSelectedFilters = selectedFilters.filter((f) => f !== label);
-        const newUnselectedFilters = [...unselectedFilters, label].sort();
-
-        setSelectedFilters(newSelectedFilters);
-        setUnselectedFilters(newUnselectedFilters);
-    };
-
-    const mergeArraysKeepingOrder = (
-        firstArray: string[],
-        secondArray: string[],
-    ): string[] => {
-        const elementsSet = new Set(firstArray);
-
-        secondArray.forEach((element) => {
-            if (!elementsSet.has(element)) {
-                firstArray.push(element);
-            }
-        });
-
-        return firstArray;
-    };
 
     useEffect(() => {
         const projectsOptions = (projects || []).map((project) => ({
@@ -171,81 +105,7 @@ export const FeatureToggleFilters: VFC<IFeatureToggleFiltersProps> = ({
         JSON.stringify(tags),
     ]);
 
-    useEffect(() => {
-        const newSelectedFilters = availableFilters
-            .filter((field) =>
-                Boolean(
-                    state[field.filterKey as keyof FeatureTogglesListFilters],
-                ),
-            )
-            .map((field) => field.label);
-        const newUnselectedFilters = availableFilters
-            .filter(
-                (field) =>
-                    !state[field.filterKey as keyof FeatureTogglesListFilters],
-            )
-            .map((field) => field.label)
-            .sort();
-
-        setSelectedFilters(
-            mergeArraysKeepingOrder(selectedFilters, newSelectedFilters),
-        );
-        setUnselectedFilters(newUnselectedFilters);
-    }, [JSON.stringify(state), JSON.stringify(availableFilters)]);
-
-    const hasAvailableFilters = unselectedFilters.length > 0;
     return (
-        <StyledBox>
-            {selectedFilters.map((selectedFilter) => {
-                const filter = availableFilters.find(
-                    (filter) => filter.label === selectedFilter,
-                );
-
-                if (!filter) {
-                    return null;
-                }
-
-                if ('dateOperators' in filter) {
-                    return (
-                        <FilterDateItem
-                            label={filter.label}
-                            state={state[filter.filterKey]}
-                            onChange={(value) =>
-                                onChange({ [filter.filterKey]: value })
-                            }
-                            operators={filter.dateOperators}
-                            onChipClose={() => deselectFilter(filter.label)}
-                        />
-                    );
-                }
-
-                return (
-                    <FilterItem
-                        key={filter.label}
-                        label={filter.label}
-                        state={state[filter.filterKey]}
-                        options={filter.options}
-                        onChange={(value) =>
-                            onChange({ [filter.filterKey]: value })
-                        }
-                        singularOperators={filter.singularOperators}
-                        pluralOperators={filter.pluralOperators}
-                        onChipClose={() => deselectFilter(filter.label)}
-                    />
-                );
-            })}
-
-            <ConditionallyRender
-                condition={hasAvailableFilters}
-                show={
-                    <AddFilterButton
-                        visibleOptions={unselectedFilters}
-                        setVisibleOptions={setUnselectedFilters}
-                        hiddenOptions={selectedFilters}
-                        setHiddenOptions={setSelectedFilters}
-                    />
-                }
-            />
-        </StyledBox>
+        <Filters availableFilters={availableFilters} state={state} onChange={onChange} />
     );
 };
